+++ qbraid/interface/qbraid_qasm3/tools.py
@@ -13,6 +13,7 @@ Module containing OpenQASM 3 tools
 
 """
 import os
+import re
 from typing import List
 
 import numpy as np
@@ -70,6 +71,127 @@ def _convert_to_contiguous_qasm3(qasmstr: QASMType, rev_qubits=False) -> QASMTyp
     circuit = loads(qasmstr)
     circuit_contig = _convert_to_contiguous_qiskit(circuit, rev_qubits=rev_qubits)
     return dumps(circuit_contig)
+
+
+def _qasm_qubit_decl(qasmstr: str) -> List[QASMType]:
+    """Get declaration statement of qasm qubits.
+
+    Args:
+        qasmstr (str): OpenQASM 2 or OpenQASM 3 string
+
+    Returns:
+        List of qubits in the circuit
+    """
+    return [
+        text.replace("\n", "")
+        for match in re.findall(r"(qreg.*;)|(qubit.*;)", qasmstr)
+        for text in match
+        if text != "" and len(text) >= 4
+    ]
+
+
+def _get_qreg(qubit_decl: str) -> tuple:
+    """Get qreg name and size from qubit declaration"""
+
+    name, size = None, None
+    if "qreg" in qubit_decl:
+        reg = qubit_decl.split("qreg")[1].split(";")[0].strip()
+        elements = reg.split("[")
+        name = elements[0]
+        size = int(elements[1].split("]")[0])
+    else:
+        try:
+            name = qubit_decl.split("]")[1].split(";")[0].strip()
+            size = int(qubit_decl.split("[")[1].split("]")[0].strip())
+        except IndexError as _:
+            name = qubit_decl.split("qubit")[1].split(";")[0].strip()
+            size = 1
+    return (name, size)
+
+
+def _remove_gate_definitions(qasm_str):
+    """This is required to account for the case when the gate
+     definition has an argument which is having same name as a
+     quantum register
+
+     now, if any gate is applied on this argument, it will be
+     interpreted as being applied on THE WHOLE register, when it is
+     only applied on the argument.
+
+    Example :
+
+    gate custom q1 {
+        x q1; // this is STILL DETECTED as a gate application on q1
+    }
+    qreg q1[4];
+    qreg q2[2];
+    custom q1[0];
+    cx q1[1], q2[1];
+
+    // Actual depth : 1
+    // Calculated depth : 2 (because of the gate definition)
+    """
+
+    gate_decls = [x.group() for x in re.finditer(r"(gate)(.*\n)*?\s*\}", qasm_str)]
+    for decl in gate_decls:
+        qasm_str = qasm_str.replace(decl, "")
+    return qasm_str
+
+
+def _get_unused_qubit_indices(qasm_str, register_list):
+    qasm_str = _remove_gate_definitions(qasm_str)
+    lines = qasm_str.splitlines()
+    gate_lines = [
+        s
+        for s in lines
+        if s.strip() and not s.strip().startswith(("OPENQASM", "include", "qreg", "qubit", "//"))
+    ]
+    unused_indices = {}
+    for qreg, size in register_list:
+        unused_indices[qreg] = set(range(size))
+
+        for line in gate_lines:
+            if qreg not in line:
+                continue
+            # either qubits or full register is referenced
+            used_indices = {int(x) for x in re.findall(rf"{qreg}\[(\d+)\]", line)}
+            if len(used_indices) > 0:
+                unused_indices[qreg] = unused_indices[qreg].difference(used_indices)
+            else:
+                # full register is referenced
+                unused_indices[qreg] = set()
+                break
+
+            if len(unused_indices[qreg]) == 0:
+                break
+
+    return unused_indices
+
+
+def convert_to_contiguous_qasm3(qasmstr: str) -> QASMType:
+    """Converts OpenQASM 3 string to contiguous qasm3 string with gate expansion"""
+
+    # SCOPE : no loops, no functions at the moment
+
+    # pylint: disable=import-outside-toplevel
+
+    # 1. Analyse the qasm3 string for qubits and bit declarations
+    qubit_list = _qasm_qubit_decl(qasmstr)
+    qreg_list = []
+
+    # 2. Identify which qubits are not used in the circuit
+    for qubit_decl in qubit_list:
+        qreg_list.append(_get_qreg(qubit_decl))
+
+    # 3. Add an identity gate for the unused qubits
+    qubit_indices = _get_unused_qubit_indices(qasmstr, qreg_list)
+    expansion_qasm = ""
+    for reg, indices in qubit_indices.items():
+        for index in indices:
+            expansion_qasm += f"i {reg}[{index}];\n"
+
+    # 4. Return the qasm3 string
+    return qasmstr + expansion_qasm
 
 
 def _build_qasm_3_reg(line: str, qreg_type: bool) -> QASMType:
